refactor: consolidate VERSION and render_version into tui.rs
Move the duplicated VERSION constant and render_version helper from
splash.rs and onboarding.rs into the shared tui.rs module. Consolidate
the two identical tests into a single test in tui::tests.
